requirements.txt
seleniumbase
groq
requests
google-generativeai